orderlines, controllers and orders complete, testing done on DAO's

+++ src/main/java/com/qa/ims/persistence/domain/Product.java
@@ -43,7 +43,7 @@ public class Product {
         this.stockQuantity = stockQuantity;
     }
 
-    public double getPrice() {
+    public Long getPrice() {
         return price;
     }
